requirements.txt
fastembed==0.3.3
flashrank==0.2.8
langchain-core>=0.2.10,<0.3.0
langsmith>=0.1.17,<0.2.0
langchain-community==0.2.6
langchain-experimental==0.0.62
langchain-groq==0.1.6
langchain-qdrant==0.1.1
langchain==0.2.6
python-dotenv==1.0.1
qdrant-client==1.10.1
streamlit==1.36.0
pypdfium2==4.30.0
huggingface-hub>=0.20,<1.0
transformers==4.35.2
tokenizers>=0.14,<0.19
# Multi-modal processing
PyMuPDF==1.23.8
pytesseract==0.3.10
Pillow>=10.3,<11.0
pdfplumber==0.10.3
sentence-transformers==3.0.0
# FastAPI
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6